refomatting post files
