+++ client/src/components/Posts/index.js
@@ -1,169 +1,124 @@
-// import React, { useEffect, useState } from "react";
-// import { useStoreContext } from "../../utils/GlobalState";
-// import { REMOVE_POST, UPDATE_POST, ADD_POST } from "../../utils/actions";
-// import { idbPromise } from "../../utils/helpers";
-// import "./posts.css";
+import React, { useState, useEffect } from "react";
+import { useLazyQuery, useMutation } from "@apollo/client";
+import {
+  QUERY_POSTS,
+  ADD_POST,
+  UPDATE_POST,
+  REMOVE_POST,
+} from "../../utils/queries";
+import { useStoreContext } from "../../utils/GlobalState";
 
-// const Posts = ({ _id, username, postBody, createdAt }) => {
-//   const [state, dispatch] = useStoreContext();
-//   const [newUsername, setNewUsername] = useState("");
-//   const [newPost, setNewPost] = useState("");
+const Posts = () => {
+  const [state, dispatch] = useStoreContext();
+  const [getPosts, { data }] = useLazyQuery(QUERY_POSTS);
+  const [addPost, { data: postData }] = useMutation(ADD_POST);
+  const [updatePost, { data: updatedPostData }] = useMutation(UPDATE_POST);
+  const [deletePost, { data: deletedPostData }] = useMutation(REMOVE_POST);
+  const [postText, setPostText] = useState("");
+  const [username, setUsername] = useState("");
 
-//   useEffect(() => {
-//     if (state.posts.length) {
-//       idbPromise("posts", "get", state.posts);
-//     }
-//   }, [state.posts]);
+  useEffect(() => {
+    getPosts();
+  }, []);
 
-//   const addPost = () => {
-//     dispatch({
-//       type: ADD_POST,
-//       payload: {
-//         username: newUsername,
-//         createdAt: new Date(),
-//         postBody: newPost,
-//       },
-//     });
-//     idbPromise("posts", "put", {
-//       username: newUsername,
-//       createdAt: new Date(),
-//       postBody: newPost,
-//     });
-//     setNewUsername("");
-//     setNewPost("");
-//   };
+  useEffect(() => {
+    if (data) {
+      dispatch({
+        type: "ADD_POST",
+        posts: data.posts,
+      });
+    }
+  }, [data]);
 
-//   const updatePost = (_id) => {
-//     const postIndex = state.posts.findIndex((post) => post._id === _id);
-//     if (postIndex !== -1) {
-//       const updatedPost = {
-//         ...state.posts[postIndex],
-//         username: newUsername,
-//         postBody: newPost,
-//       };
-//       const updatedPosts = [
-//         ...state.posts.slice(0, postIndex),
-//         updatedPost,
-//         ...state.posts.slice(postIndex + 1),
-//       ];
-//       dispatch({
-//         type: UPDATE_POST,
-//         posts: updatedPosts,
-//       });
-//       idbPromise("posts", "put", updatedPosts);
-//     }
-//   };
+  useEffect(() => {
+    console.log(postText);
+  }, [postText]);
 
-//   const removePost = (_id) => {
-//     const postIndex = state.posts.findIndex((post) => post._id === _id);
-//     if (postIndex !== -1) {
-//       const updatedPosts = [
-//         ...state.posts.slice(0, postIndex),
-//         ...state.posts.slice(postIndex + 1),
-//       ];
-//       dispatch({
-//         type: REMOVE_POST,
-//         posts: updatedPosts,
-//       });
-//       idbPromise("posts", "put", updatedPosts);
-//     }
-//   };
+  useEffect(() => {
+    console.log(username);
+  }, [username]);
 
-//   return (
-//     <div>
-//       <div className="card px-1 py-1">
-//         <p className="m-0">
-//           {username} posted on {createdAt}
-//         </p>
-//         <p>{postBody}</p>
-//         <button onClick={() => updatePost(_id)}>Update this post</button>
-//         <button onClick={() => removePost(_id)}>Remove this post</button>
-//       </div>
-//       <hr />
-//       <div className="add-post">
-//         <input
-//           type="text"
-//           placeholder="New post"
-//           value={newPost}
-//           onChange={(e) => setNewPost(e.target.value)}
-//         />
-//         <button onClick={addPost}>Add Post</button>
-//       </div>
-//     </div>
-//   );
-// };
+  const handleSubmit = (e) => {
+    console.log("submit handled");
+    e.preventDefault();
+    addPost({
+      variables: { postText, username },
+      update: (store, { data: { addPost } }) => {
+        const data = store.readQuery({ query: QUERY_POSTS });
+        store.writeQuery({
+          query: QUERY_POSTS,
+          data: { posts: [...data.posts, addPost] },
+        });
+      },
+    });
+    setPostText("");
+    setUsername("");
+  };
 
-// export default Posts;
+  const handleUpdate = (e, postId) => {
+    e.preventDefault();
+    updatePost({
+      variables: { postId, postText, username },
+      update: (store, { data: { updatePost } }) => {
+        const data = store.readQuery({ query: QUERY_POSTS });
+        store.writeQuery({
+          query: QUERY_POSTS,
+          data: {
+            posts: data.posts.map((post) =>
+              post._id === postId ? updatePost : post
+            ),
+          },
+        });
+      },
+    });
+    setPostText("");
+    setUsername("");
+  };
 
-// import React, { useState, useEffect } from "react";
-// import { useLazyQuery, useMutation } from "@apollo/client";
-// import { QUERY_POSTS, ADD_POST } from "../../utils/queries";
-// import { useStoreContext } from "../../utils/GlobalState";
+  const handleDelete = (e, postId) => {
+    e.preventDefault();
+    deletePost({
+      variables: { postId },
+      update: (store, { data: { deletePost } }) => {
+        const data = store.readQuery({ query: QUERY_POSTS });
+        store.writeQuery({
+          query: QUERY_POSTS,
+          data: {
+            posts: data.posts.filter((post) => post._id !== postId),
+          },
+        });
+      },
+    });
+  };
 
-// const Posts = () => {
-//   const [state, dispatch] = useStoreContext();
-//   const [getPosts, { data }] = useLazyQuery(QUERY_POSTS);
-//   const [addPost, { data: postData }] = useMutation(ADD_POST);
-//   const [postText, setPostText] = useState("");
-//   const [username, setUsername] = useState("");
+  return (
+    <div>
+      <form onSubmit={handleSubmit}>
+        <input
+          type="text"
+          placeholder="Username"
+          value={username}
+          onChange={(e) => setUsername(e.target.value)}
+        />
+        <input
+          type="text"
+          placeholder="Post"
+          value={postText}
+          onChange={(e) => setPostText(e.target.value)}
+        />
+        <button type="submit">Post</button>
+      </form>
+      {state.posts.map((post) => (
+        <div key={post._id}>
+          <p>
+            <b>{post.username}</b> {post.postText}
+          </p>
+          <button onClick={(e) => handleUpdate(e, post._id)}>Update</button>
+          <button onClick={(e) => handleDelete(e, post._id)}>Delete</button>
+        </div>
+      ))}
+    </div>
+  );
+};
 
-//   useEffect(() => {
-//     getPosts();
-//   }, []);
-
-//   useEffect(() => {
-//     if (data) {
-//       dispatch({
-//         type: "SET_POSTS",
-//         posts: data.posts,
-//       });
-//     }
-//   }, [data]);
-
-//   const handleSubmit = (e) => {
-//     e.preventDefault();
-//     addPost({
-//       variables: { postText, username },
-//       update: (store, { data: { addPost } }) => {
-//         const data = store.readQuery({ query: QUERY_POSTS });
-//         store.writeQuery({
-//           query: QUERY_POSTS,
-//           data: { posts: [...data.posts, addPost] },
-//         });
-//       },
-//     });
-//     setPostText("");
-//     setUsername("");
-//   };
-
-//   return (
-//     <div>
-//       {state.posts.map((post) => (
-//         <div key={post._id} className="card px-1 py-1">
-//           <p className="m-0">
-//             {post.username} posted on {post.createdAt}
-//           </p>
-//           <p>{post.postBody}</p>
-//         </div>
-//       ))}
-//       <hr />
-//       <form onSubmit={handleSubmit}>
-//         <input
-//           type="text"
-//           placeholder="Username"
-//           value={username}
-//           onChange={(e) => setUsername(e.target.value)}
-//         />
-//         <input
-//           type="text"
-//           placeholder="New post"
-//           value={postText}
-//           onChange={(e) => setPostText(e.target.value)}
-//         />
-//         <button type="submit">Add Post</button>
-//       </form>
-//     </div>
-//   );
-// };
-
-// export default Posts;
+export default Posts;
