+++ client/src/pages/Home.js
@@ -3,7 +3,7 @@ import ProductList from "../components/ProductList";
 import Cart from "../components/Cart";
 import Title from "../components/Title";
 import Band from "../components/Band";
-import PostList from "../components/PostList";
+import Posts from "../components/Posts";
 
 
 const Home = () => {
@@ -13,7 +13,7 @@ const Home = () => {
       <Band />
       {/* <Video /> */}
       <ProductList />
-      <PostList />
+      <Posts />
       <Cart />
     </div>
   );
